rename myGui.h/.cpp to myUi.h/.cpp, add my::ImeEditBox
git-svn-id: svn://127.0.0.1/repos/trunk@137 a5c22d9f-9f17-4845-9224-d5bed68bb274

+++ myd3dlib/include/myFont.h
@@ -185,7 +185,7 @@ namespace my
 		static FontPtr CreateFontFromFile(
 			LPDIRECT3DDEVICE9 pDevice,
 			LPCSTR pFilename,
-			float height,
+			int height,
 			unsigned short pixel_gap = 0,
 			FT_Long face_index = 0);
 
@@ -193,7 +193,7 @@ namespace my
 			LPDIRECT3DDEVICE9 pDevice,
 			const void * file_base,
 			long file_size,
-			float height,
+			int height,
 			unsigned short pixel_gap = 0,
 			FT_Long face_index = 0);
